Rename Mixer->EntityMixer and move to sound module
